# docs/openapi.yaml
openapi: 3.1.0
info:
  title: IP Intelligence Service API
  version: 1.0.0
  description: Private IP intelligence API shared by multiple navigation sites.
servers:
  - url: https://ip-intelligence.example.com
    description: Placeholder production endpoint
paths:
  /api/public/lookup:
    get:
      summary: Resolve one IP for the public query interface
      parameters:
        - in: query
          name: ip
          required: false
          schema: { type: string, maxLength: 64 }
          description: IPv4 or IPv6 input; omitted means the current connection address
      responses:
        '200':
          description: Single IP processed without persisting the queried address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicLookupResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceNotReady' }
  /health:
    get:
      summary: Process liveness check
      responses:
        '200':
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /ready:
    get:
      summary: Query readiness check
      security:
        - ClientId: []
          Timestamp: []
          Nonce: []
          Signature: []
      responses:
        '200':
          description: Required providers are ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadyResponse'
        '503':
          $ref: '#/components/responses/ServiceNotReady'
  /metrics:
    get:
      summary: Prometheus metrics for this service instance
      security:
        - MetricsBearer: []
      responses:
        '200':
          description: Prometheus text exposition
          content:
            text/plain:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/ip/lookup:
    post:
      summary: Resolve one to one hundred IP inputs
      security:
        - ClientId: []
          Timestamp: []
          Nonce: []
          Signature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LookupRequest'
      responses:
        '200':
          description: Batch processed; inspect each item status
          headers:
            X-Request-ID: { schema: { type: string } }
            X-Rate-Limit-Limit: { schema: { type: integer } }
            X-Rate-Limit-Remaining: { schema: { type: integer } }
            X-Rate-Limit-Reset: { schema: { type: integer }, description: Seconds until the current minute window resets }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LookupResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/ReplayDetected' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/ServiceNotReady' }
  /v1/meta/sources:
    get:
      summary: List provider versions and freshness
      security:
        - ClientId: []
          Timestamp: []
          Nonce: []
          Signature: []
      responses:
        '200':
          description: Provider metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourcesResponse'
  /v1/meta/observability:
    get:
      summary: Aggregated runtime observability snapshot
      security:
        - ClientId: []
          Timestamp: []
          Nonce: []
          Signature: []
      responses:
        '200':
          description: Current instance runtime state without request or IP details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservabilityResponse'
components:
  securitySchemes:
    ClientId: { type: apiKey, in: header, name: X-Client-ID }
    Timestamp: { type: apiKey, in: header, name: X-Timestamp }
    Nonce: { type: apiKey, in: header, name: X-Nonce }
    Signature: { type: apiKey, in: header, name: X-Signature }
    MetricsBearer: { type: http, scheme: bearer }
  schemas:
    PublicLookupResponse:
      type: object
      required: [request_id, code, data, meta]
      properties:
        request_id: { type: string }
        code: { type: string, const: OK }
        data: { $ref: '#/components/schemas/LookupItem' }
        meta: { $ref: '#/components/schemas/LookupMeta' }
    HealthResponse:
      type: object
      required: [request_id, ok, service]
      properties:
        request_id: { type: string }
        ok: { type: boolean, const: true }
        service: { type: string, const: ip-intelligence-service }
    ReadyResponse:
      type: object
      required: [request_id, ok, required_sources_ready, sources]
      properties:
        request_id: { type: string }
        ok: { type: boolean }
        required_sources_ready: { type: boolean }
        sources:
          type: array
          items: { $ref: '#/components/schemas/SourceStatus' }
    LookupRequest:
      type: object
      additionalProperties: false
      required: [ips]
      properties:
        ips:
          type: array
          minItems: 1
          maxItems: 100
          items: { type: string, maxLength: 64 }
    LookupResponse:
      type: object
      required: [request_id, code, data, meta]
      properties:
        request_id: { type: string }
        code: { type: string, const: OK }
        data:
          type: array
          items: { $ref: '#/components/schemas/LookupItem' }
        meta: { $ref: '#/components/schemas/LookupMeta' }
    LookupItem:
      type: object
      required: [input, status, message, ip, ip_version, scope, country_code, country_name, region, city, asn, asn_org, isp, network_type, is_mobile, is_hosting, is_proxy, is_vpn, is_tor, is_anycast, confidence, evidence, sources]
      properties:
        input: { type: string }
        status: { type: string, enum: [resolved, invalid, unavailable] }
        message: { type: [string, 'null'] }
        ip: { type: [string, 'null'] }
        ip_version: { type: [integer, 'null'], enum: [4, 6, null] }
        scope: { type: string, enum: [public, private, loopback, link_local, multicast, reserved, unknown] }
        country_code: { type: [string, 'null'] }
        country_name: { type: [string, 'null'] }
        region: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        asn: { type: [integer, 'null'], minimum: 0 }
        asn_org: { type: [string, 'null'] }
        isp: { type: [string, 'null'] }
        network_type: { type: string, enum: [residential, mobile, business, education, government, hosting, cdn, unknown] }
        is_mobile: { type: [boolean, 'null'] }
        is_hosting: { type: [boolean, 'null'] }
        is_proxy: { type: [boolean, 'null'] }
        is_vpn: { type: [boolean, 'null'] }
        is_tor: { type: [boolean, 'null'] }
        is_anycast: { type: [boolean, 'null'] }
        confidence: { type: string, enum: [unknown, low, medium, high] }
        evidence:
          type: array
          items: { $ref: '#/components/schemas/Evidence' }
        sources:
          type: array
          items: { type: string }
    Evidence:
      type: object
      additionalProperties: false
      required: [source, field, value, confidence]
      properties:
        source: { type: string }
        field: { type: string }
        value: { type: string }
        confidence: { type: string, enum: [low, medium, high] }
    LookupMeta:
      type: object
      required: [requested_count, unique_count, resolved_count, invalid_count, unavailable_count, generated_at, database_versions]
      properties:
        requested_count: { type: integer, minimum: 1, maximum: 100 }
        unique_count: { type: integer, minimum: 1, maximum: 100 }
        resolved_count: { type: integer, minimum: 0 }
        invalid_count: { type: integer, minimum: 0 }
        unavailable_count: { type: integer, minimum: 0 }
        generated_at: { type: string, format: date-time }
        database_versions:
          type: object
          additionalProperties: { type: string }
    SourceStatus:
      type: object
      required: [id, status, required, version, updated_at, expires_at, last_error]
      properties:
        id: { type: string }
        status: { type: string, enum: [ready, stale, unavailable] }
        required: { type: boolean }
        version: { type: [string, 'null'] }
        updated_at: { type: [string, 'null'], format: date-time }
        expires_at: { type: [string, 'null'], format: date-time }
        last_error: { type: [string, 'null'] }
    SourcesResponse:
      type: object
      required: [request_id, code, data]
      properties:
        request_id: { type: string }
        code: { type: string, const: OK }
        data:
          type: array
          items: { $ref: '#/components/schemas/SourceStatus' }
    ObservabilityResponse:
      type: object
      required: [request_id, code, data]
      properties:
        request_id: { type: string }
        code: { type: string, const: OK }
        data:
          type: object
          description: Aggregate process, request, lookup, update, pool and readiness state.
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [request_id, code, message, details]
      properties:
        request_id: { type: string }
        code: { type: string }
        message: { type: string }
        details: {}
  responses:
    BadRequest:
      description: Invalid JSON or request structure
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Unauthorized:
      description: Missing or invalid authentication
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    NotFound:
      description: Route is disabled or does not exist
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Forbidden:
      description: Client disabled
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    ReplayDetected:
      description: Nonce was already used
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    MethodNotAllowed:
      description: Route exists but does not support the request method
      headers:
        Allow: { schema: { type: string } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    PayloadTooLarge:
      description: Request body too large
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    UnsupportedMediaType:
      description: Lookup requests must use application/json
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Unprocessable:
      description: Batch exceeds the contract limit
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    RateLimited:
      description: Client rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
        X-Rate-Limit-Limit: { schema: { type: integer } }
        X-Rate-Limit-Remaining: { schema: { type: integer, const: 0 } }
        X-Rate-Limit-Reset: { schema: { type: integer } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    InternalError:
      description: Internal server error
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    ServiceNotReady:
      description: Required data sources are unavailable
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }